Extract reifyLayer to a new file
