+++ app/javascript/projects/project_editor.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import './project_editor.css'
 import { LayerPalette } from './layer_palette'
 import { Toolbar } from './toolbar'
-import { reifyLayer } from './layers'
+import { reifyLayer } from './reify_layer'
 import { Project } from './project'
 import { Sidebar, CollapsedSidebar } from './sidebar'
 import { MapView } from './map_view'
